refactor(combinatorics): isolate counting contracts
